Extract non-C# symbol name normalizers
